+++ app/assets/javascripts/i18n.js
@@ -70,20 +70,24 @@
   var isArray = function(val) {
     if (Array.isArray) {
       return Array.isArray(val);
-    };
+    }
     return Object.prototype.toString.call(val) === '[object Array]';
   };
 
   var isString = function(val) {
-    return typeof value == 'string' || Object.prototype.toString.call(val) === '[object String]';
+    return typeof val === 'string' || Object.prototype.toString.call(val) === '[object String]';
   };
 
   var isNumber = function(val) {
-    return typeof val == 'number' || Object.prototype.toString.call(val) === '[object Number]';
+    return typeof val === 'number' || Object.prototype.toString.call(val) === '[object Number]';
   };
 
   var isBoolean = function(val) {
     return val === true || val === false;
+  };
+
+  var isNull = function(val) {
+    return val === null;
   };
 
   var decimalAdjust = function(type, value, exp) {
@@ -103,7 +107,7 @@
     // Shift back
     value = value.toString().split('e');
     return +(value[0] + 'e' + (value[1] ? (+value[1] + exp) : exp));
-  }
+  };
 
   var lazyEvaluate = function(message, scope) {
     if (isFunction(message)) {
@@ -111,13 +115,13 @@
     } else {
       return message;
     }
-  }
+  };
 
   var merge = function (dest, obj) {
     var key, value;
     for (key in obj) if (obj.hasOwnProperty(key)) {
       value = obj[key];
-      if (isString(value) || isNumber(value) || isBoolean(value)) {
+      if (isString(value) || isNumber(value) || isBoolean(value) || isArray(value) || isNull(value)) {
         dest[key] = value;
       } else {
         if (dest[key] == null) dest[key] = {};
@@ -309,11 +313,11 @@
       var firstFallback = null;
       var secondFallback = null;
       if (localeParts.length === 3) {
-        firstFallback = localeParts[0];
-        secondFallback = [
+        firstFallback = [
           localeParts[0],
           localeParts[1]
         ].join("-");
+        secondFallback = localeParts[0];
       }
       else if (localeParts.length === 2) {
         firstFallback = localeParts[0];
@@ -385,10 +389,9 @@
   // This is used internally by some functions and should not be used as an
   // public API.
   I18n.lookup = function(scope, options) {
-    options = options || {}
+    options = options || {};
 
     var locales = this.locales.get(options.locale).slice()
-      , requestedLocale = locales[0]
       , locale
       , scopes
       , fullScope
@@ -445,9 +448,8 @@
 
   // Lookup dedicated to pluralization
   I18n.pluralizationLookup = function(count, scope, options) {
-    options = options || {}
+    options = options || {};
     var locales = this.locales.get(options.locale).slice()
-      , requestedLocale = locales[0]
       , locale
       , scopes
       , translations
@@ -469,16 +471,16 @@
         if (!isObject(translations)) {
           break;
         }
-        if (scopes.length == 0) {
+        if (scopes.length === 0) {
           message = this.pluralizationLookupWithoutFallback(count, locale, translations);
         }
       }
-      if (message != null && message != undefined) {
+      if (typeof message !== "undefined" && message !== null) {
         break;
       }
     }
 
-    if (message == null || message == undefined) {
+    if (typeof message === "undefined" || message === null) {
       if (isSet(options.defaultValue)) {
         if (isObject(options.defaultValue)) {
           message = this.pluralizationLookupWithoutFallback(count, options.locale, options.defaultValue);
@@ -568,11 +570,12 @@
 
   // Translate the given scope with the provided options.
   I18n.translate = function(scope, options) {
-    options = options || {}
+    options = options || {};
 
     var translationOptions = this.createTranslationOptions(scope, options);
 
     var translation;
+    var usedScope = scope;
 
     var optionsWithoutDefault = this.prepareOptions(options)
     delete optionsWithoutDefault.defaultValue
@@ -582,7 +585,8 @@
     var translationFound =
       translationOptions.some(function(translationOption) {
         if (isSet(translationOption.scope)) {
-          translation = this.lookup(translationOption.scope, optionsWithoutDefault);
+          usedScope = translationOption.scope;
+          translation = this.lookup(usedScope, optionsWithoutDefault);
         } else if (isSet(translationOption.message)) {
           translation = lazyEvaluate(translationOption.message, scope);
         }
@@ -598,8 +602,12 @@
 
     if (typeof(translation) === "string") {
       translation = this.interpolate(translation, options);
+    } else if (isArray(translation)) {
+      translation = translation.map(function(t) {
+        return (typeof(t) === "string" ? this.interpolate(t, options) : t);
+      }, this);
     } else if (isObject(translation) && isSet(options.count)) {
-      translation = this.pluralize(options.count, scope, options);
+      translation = this.pluralize(options.count, usedScope, options);
     }
 
     return translation;
@@ -607,7 +615,11 @@
 
   // This function interpolates the all variables in the given message.
   I18n.interpolate = function(message, options) {
-    options = options || {}
+    if (message === null) {
+      return message;
+    }
+
+    options = options || {};
     var matches = message.match(this.placeholder)
       , placeholder
       , value
@@ -618,8 +630,6 @@
     if (!matches) {
       return message;
     }
-
-    var value;
 
     while (matches.length) {
       placeholder = matches.shift();
@@ -633,7 +643,7 @@
         value = this.missingPlaceholder(placeholder, message, options);
       }
 
-      regex = new RegExp(placeholder.replace(/\{/gm, "\\{").replace(/\}/gm, "\\}"));
+      regex = new RegExp(placeholder.replace(/{/gm, "\\{").replace(/}/gm, "\\}"));
       message = message.replace(regex, value);
     }
 
@@ -645,14 +655,14 @@
   // which will be retrieved from `options`.
   I18n.pluralize = function(count, scope, options) {
     options = this.prepareOptions({count: String(count)}, options)
-    var pluralizer, message, result;
+    var pluralizer, result;
 
     result = this.pluralizationLookup(count, scope, options);
-    if (result.translations == undefined || result.translations == null) {
+    if (typeof result.translations === "undefined" || result.translations == null) {
       return this.missingTranslation(scope, options);
     }
 
-    if (result.message != undefined && result.message != null) {
+    if (typeof result.message !== "undefined" && result.message != null) {
       return this.interpolate(result.message, options);
     }
     else {
@@ -664,7 +674,7 @@
   // Return a missing translation message for the given parameters.
   I18n.missingTranslation = function(scope, options) {
     //guess intended string
-    if(this.missingBehaviour == 'guess'){
+    if(this.missingBehaviour === 'guess'){
       //get only the last portion of the scope
       var s = scope.split('.').slice(-1)[0];
       //replace underscore with space && camelcase with space and lowercase letter
@@ -826,7 +836,7 @@
     // we have a date, so just return it.
     if (typeof(date) == "object") {
       return date;
-    };
+    }
 
     matches = date.toString().match(/(\d{4})-(\d{2})-(\d{2})(?:[ T](\d{2}):(\d{2}):(\d{2})([\.,]\d{1,3})?)?(Z|\+00:?00)?/);
 
@@ -1024,7 +1034,7 @@
   };
 
   I18n.getFullScope = function(scope, options) {
-    options = options || {}
+    options = options || {};
 
     // Deal with the scope as an array.
     if (isArray(scope)) {